Made 100-atoi.c follows betty style

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -29,7 +29,7 @@ int _atoi(char *s)
 			{
 				return (2147483647);  /* Return INT_MAX on overflow */
 			}
-			else if (sign == -1 && result > (2147483648U - (*s - '0')) / 10)
+			else if (sign == -1 && result > (int)(2147483648U - (*s - '0')) / 10)
 			{
 				return (-2147483648); /* Return INT_MIN on overflow */
 			}
